Move RevertBatches function to new file and unlink its test suite
